# openapi.yaml
components:
  schemas:
    BinaryFileBodyModel:
      properties:
        fileType:
          description: The file type that should be used to build the binary file
            into a SCMF
          title: Filetype
          type: integer
        overwrite:
          default: true
          description: if True will overwrite any existing onboard file
          title: Overwrite
          type: boolean
        sessionId:
          description: The AMPCS session to send the command through
          title: Sessionid
          type: integer
        sourceFilePath:
          description: 'The full path on GDS host (ex: /proj/europa/sit/current/files/myfile.data)'
          title: Sourcefilepath
          type: string
        stringSelection:
          allOf:
          - $ref: '#/components/schemas/StringSelection'
          default: DEFAULT
          description: Specifies the MTAK string to send the command
        targetFilePath:
          description: The full onboard path (i.e. /eng1/myfile.data
          title: Targetfilepath
          type: string
        timeout:
          default: 10
          description: The timeout on the dispatch process in seconds
          minimum: 0.0
          title: Timeout
          type: integer
      required:
      - sessionId
      - sourceFilePath
      - targetFilePath
      - fileType
      title: BinaryFileBodyModel
      type: object
...
    CmdDispatchedResp:
      properties:
        cmdRequested:
          default: ''
          description: This mirrors what was received as the requested command string
          title: Cmdrequested
          type: string
        dispatchTime:
          default: ''
          description: 'Time that the dispatch occurred. Example: 2017-09-25T03:57:42.676Z'
          title: Dispatchtime
          type: string
      title: CmdDispatchedResp
      type: object
...
    ErrorResponse:
      properties:
        message:
          default: ''
          description: Error message
          title: Message
          type: string
      title: ErrorResponse
      type: object
...
    StringSelection:
      description: An enumeration.
      enum:
      - DEFAULT
      - A
      - B
      - AB
      title: StringSelection
      type: string
...
info:
  description: RESTful API of Ingenium VenueServer that interfaces with test venues
  title: Ingenium VenueServer
  version: 14.1-TP
openapi: 3.1.0
paths:
  /api/v3/cmd/binary_file:
    post:
      operationId: binary_file_api_v3_cmd_binary_file_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BinaryFileBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CmdDispatchedResp'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Upload a file to flight computer via MTAK
      tags:
      - COMMAND
  /api/v3/cmd/fsw_cmd:
    post:
      operationId: fsw_cmd_api_v3_cmd_fsw_cmd_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FswCmdBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CmdDispatchedResp'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Send a FSW command via MTAK
      tags:
      - COMMAND
  /api/v3/cmd/hw_cmd:
    post:
      operationId: hw_cmd_api_v3_cmd_hw_cmd_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HwCmdBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CmdDispatchedResp'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Send a HW command via MTAK
      tags:
      - COMMAND
  /api/v3/cmd/scmf:
    post:
      operationId: scmf_file_api_v3_cmd_scmf_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScmfFileBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CmdDispatchedResp'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Upload a SCMF file to flight
      tags:
      - COMMAND
  /api/v3/cmd/sse:
    post:
      description: Send a SSE (Simulation and Support Equipment) command via MTAK
      operationId: sse_cmd_api_v3_cmd_sse_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SseCmdBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CmdDispatchedResp'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Send a SSE command via MTAK
      tags:
      - COMMAND
  /api/v3/eha/realtime:
    get:
      description: Queries AMPCS Global LAD and return an array of EHA channel objects
      operationId: eha_realtime_api_v3_eha_realtime_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EhaRtBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ChannelValueObjectRespModel'
                title: Response 200 Eha Realtime Api V3 Eha Realtime Get
                type: array
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Queries the real time telemetry stream for a given channel
      tags:
      - EHA
  /api/v3/eha/realtime_multi:
    get:
      description: Queries AMPCS Global LAD and return an array of EHA channel objects
      operationId: eha_realtime_multi_api_v3_eha_realtime_multi_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EhaRtMultiBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ChannelValueObjectRespModel'
                title: Response 200 Eha Realtime Multi Api V3 Eha Realtime Multi Get
                type: array
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Queries the real time telemetry stream for a given channel. This supports
        querying multiple channel ids, etc.
      tags:
      - EHA
  /api/v3/evr/realtime:
    get:
      description: Queries the AMPCS Global LAD and return an array of EVR objects
      operationId: evr_realtime_api_v3_evr_realtime_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvrRtBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EVRObjectResp'
                title: Response 200 Evr Realtime Api V3 Evr Realtime Get
                type: array
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Queries the real time EVR stream for the specified EVRs
      tags:
      - EVR
...
  /api/v3/mtak/shutdown:
    post:
      operationId: shutdown_mtak_api_v3_mtak_shutdown_post
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Shutdown MTAK that was started by VenueServer
      tags:
      - MTAK
...